# openapi/transaction-verification-internal.yaml
openapi: 3.0.3
info:
  title: 'Transaction Verification Internal API (service-to-service)'
  version: '1.0'
servers:
  -
    url: 'https://{host}/api/internal/transaction-verification/v1'
    variables:
      host:
        default: api.example.com
        description: 'API host — override for your environment (staging vs production).'
security:
  -
    hmacService: []
paths:
...
  '/verifications/{uuid}/reprocess':
    post:
      operationId: transaction-verification.internal.v1.verifications.reprocess
      summary: 'Read the receipt again with the current engines and decide again'
      tags:
        - Verification
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        202:
          description: 'Queued again; a receipt still waiting or being read is returned as it is.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      uuid:
                        type: string
                      subject_type:
                        type: string
                      subject_id:
                        type: string
                      status:
                        type: string
                      verdict:
                        type: string
                        nullable: true
                      checks:
                        type: object
                        additionalProperties:
                          type: object
                          additionalProperties: {  }
                      duplicate_of:
                        type: array
                        items:
                          type: string
                      extracted:
                        type: object
                        additionalProperties:
                          anyOf:
                            -
                              type: string
                            -
                              type: integer
                          nullable: true
                      error:
                        type: string
                        nullable: true
                    required:
                      - uuid
                      - subject_type
                      - subject_id
                      - status
                      - verdict
                      - checks
                      - duplicate_of
                      - extracted
                      - error
                required:
                  - success
                  - data
        404:
          description: 'No verification with this uuid.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                required:
                  - success
                  - error
                  - message
        401:
          description: 'Missing, wrong, expired or reused signature.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                required:
                  - success
                  - error
                  - message
        403:
          description: 'Caller IP not allowed.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                required:
                  - success
                  - error
                  - message
        429:
          description: 'More than 30 writes a minute.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                required:
                  - success
                  - error
                  - message
        503:
          description: 'The API is switched off.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                required:
                  - success
                  - error
                  - message
...
components:
  securitySchemes:
    hmacService:
      type: apiKey
      description: "HMAC-SHA256 service authentication. Send ALSO (all REQUIRED):\nX-Service-Key-Id, X-Service-Timestamp, X-Service-Nonce.\n\ncanonical = METHOD \\n PATH?SORTED_QUERY \\n sha256hex(rawBody) \\n TIMESTAMP \\n NONCE\nsignature = lowercase-hex HMAC-SHA256(canonical, secret)\nConstraints: timestamp within ±300s of now; nonce = 32 lowercase hex, single-use.\n\nWorked vector — verify your signer reproduces this exactly BEFORE calling the API.\nThis uses a TEST key/secret (\"test_secret_do_not_use\") — NEVER a production credential.\n  secret : test_secret_do_not_use\n  request: GET /api/internal/affiliate/v1/affiliates/abc-123/wallet?merchant_id=11&page=2\n  canonical (\\n-joined lines):\n    GET\n    /api/internal/affiliate/v1/affiliates/abc-123/wallet?merchant_id=11&page=2\n    e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855   # sha256(\"\")\n    1700000000\n    0123456789abcdef0123456789abcdef\n  X-Service-Signature = ad09764034d2dcbf9d27a3ab26ace242390f7c535ef4d995485a9128a2631755"
      in: header
      name: X-Service-Signature